+++ pvo_low_io_single.h
@@ -56,13 +56,13 @@ int pvo_low_io_single_close( pvo_low_io_file_handle_t self );
 /// for more information.
 int pvo_low_io_single_write_single( pvo_low_io_file_handle_t self,
                                     void* buf,
-                                    int   count );
+                                    size_t count );
 
 /// Write ordered. See the documentation in pvo_low_io.h
 /// for more information.
 int pvo_low_io_single_write_ordered( pvo_low_io_file_handle_t self,
-                                     void* buf,
-                                     int   count );
+                                     void*  buf,
+                                     size_t count );
 
 #ifdef __cplusplus
 }
